Fix a bug with sidebar hovering (#30)
This will fix an issue where hovering next to the sidebar triggered some of the
events. I had to make the whole sidebar, not just the container that contains
its contents, smaller and expand that on hover.

This comes with a caveat: the animations that happens inside TCR (like slowly
fading in the names of the tabs and hidind the settings icon) don't work because
they are not just hidden away by overflow now but, because the size of the whole
box changes, they would wrap around. So they have to become invisible
immediately and not after a delay. The animations therefore looks a bit less
smooth now.