Windows: make it possible to drag the maximized window by its top ledge.
In Edge and stock Firefox, there is that _one_ offset pixel on the top which you could use to drag the whole window around when it is maximized (without affecting neither rightmost close pixel nor leftmost back button pixel). This adds a minimally visually intrusive half-pixel ledge which makes this behavior possible as well. Back, forward, reload, Overflow and hamburger buttons need to be offset back half a pixel to make them be responsive to uppermost ledge - an important convenience function as it reduces a degree of precision required to access these buttons.